-Added support for putting multiple messages in the same corner above or below each other. -Added "Pause" OSD message. Had to move the rompause code in interupt.c lower so that the OSD message would be rendered on pause.

+++ main/osd.h
@@ -73,14 +73,14 @@ enum osd_animation_type {
 typedef struct {
     char *text; // Text that this object will have when displayed
     enum osd_corner corner; // One of the 9 corners
-    int xoffset; // Relative X position
-    int yoffset; // Relative Y position
+    float xoffset; // Relative X position
+    float yoffset; // Relative Y position
     float color[3]; // Red, Green, Blue values
     int state; // display state of current message
     enum osd_animation_type animation[OSD_NUM_STATES]; // animations for each display state
-    int timeout[OSD_NUM_STATES]; // timeouts for each display state
+    unsigned int timeout[OSD_NUM_STATES]; // timeouts for each display state
 #define OSD_INFINITE_TIMEOUT 0xffffffff
-    int frames; // number of frames in this state
+    unsigned int frames; // number of frames in this state
 } osd_message_t;
 
 enum { R, G, B }; // for referencing color array
